requirements.txt
fastapi
uvicorn
motor
pymongo
python-dotenv
jinja2
pydantic
pydantic-core
